Move model tests into test_models.py
